Make QR code scanner into a component
