+++ app/(tabs)/(addPrescriptions)/QRCodeScreen.tsx
@@ -1,21 +1,18 @@
 import { useAgent } from "@credo-ts/react-hooks";
 import { useQueryClient } from "@tanstack/react-query";
-import { CameraView, useCameraPermissions } from "expo-camera/next";
 import { useRouter } from "expo-router";
 import { useEffect, useState } from "react";
-import { View, Text, StyleSheet, Pressable } from "react-native";
 
 import {
   resolveAndGetCredentialsWithAgent,
   storeIssuerNameFromOfferWithAgent,
 } from "@/agent/Vc";
 import LoadingComponent from "@/components/LoadingComponent";
+import QRCodeScannerComponent from "@/components/QRCodeScannerComponent";
 import { useCredentialResponsesStore } from "@/state/CredentialResponsesStore";
 import { useIssuerInfoStore } from "@/state/IssuerInfoStore";
-import { defaultStyles } from "@/stylesheets/DefaultStyles";
 
 export default function QRCodeScreen() {
-  const [permission, requestPermission] = useCameraPermissions();
   const [scannedData, setScannedData] = useState<string>("");
   const [receivingState, setReceivingState] = useState(false);
   const router = useRouter();
@@ -24,8 +21,6 @@ export default function QRCodeScreen() {
 
   const setCredentialResponses = useCredentialResponsesStore(({ set }) => set);
   const setIssuerInfo = useIssuerInfoStore(({ set }) => set);
-
-  const isPermissionGranted = Boolean(permission?.granted);
 
   useEffect(() => {
     const handleUpload = async () => {
@@ -72,56 +67,7 @@ export default function QRCodeScreen() {
 
   if (receivingState) return <LoadingComponent />;
 
-  if (!permission) {
-    return <Text>Requesting camera permission...</Text>;
-  }
-
   return (
-    <View style={defaultStyles.container}>
-      <View style={defaultStyles.overlay}>
-        <Text style={defaultStyles.overlayText}>
-          Scan QR code {"\n"}Place the QR code in the camera's field of view
-        </Text>
-      </View>
-
-      {!isPermissionGranted ? (
-        <Pressable style={styles.button} onPress={requestPermission}>
-          <Text style={styles.buttonText}>Grant Camera Permission</Text>
-        </Pressable>
-      ) : (
-        <CameraView
-          style={styles.camera}
-          facing="back"
-          onBarcodeScanned={({ data }) => {
-            setScannedData(data);
-            console.log("Scanned Data:", data);
-          }}
-        />
-      )}
-      {scannedData && (
-        <Text style={styles.result}>Scanned url: {scannedData}</Text>
-      )}
-    </View>
+      <QRCodeScannerComponent onScan={setScannedData} />
   );
 }
-
-const styles = StyleSheet.create({
-  camera: {
-    width: "100%",
-    height: "70%",
-  },
-  button: {
-    padding: 12,
-    backgroundColor: "#007bff",
-    borderRadius: 5,
-  },
-  buttonText: {
-    color: "#fff",
-    fontWeight: "bold",
-  },
-  result: {
-    marginTop: 10,
-    fontSize: 16,
-    color: "#333",
-  },
-});
